Database.SQLite3

Connection management

withDatabase

Arguments

:: Text
connection string

-> (Database -> IO a)
user program

-> IO a

Manage a connection to a database in an exception-safe way (with bracket)

open :: Text -> IO Database

https://www.sqlite.org/c3ref/open.html

open2 :: Text -> [SQLOpenFlag] -> SQLVFS -> IO Database

https://www.sqlite.org/c3ref/open.html

close :: Database -> IO ()

https://www.sqlite.org/c3ref/close.html

Simple query execution

https://sqlite.org/c3ref/exec.html

exec :: Database -> Text -> IO ()

Execute zero or more SQL statements delimited by semicolons.

execPrint :: Database -> Text -> IO ()

Like exec, but print result rows to stdout.

This is mainly for convenience when experimenting in GHCi. The output format may change in the future.

execWithCallback :: Database -> Text -> ExecCallback -> IO ()

Like exec, but invoke the callback for each result row.

type ExecCallback

Arguments

= ColumnCount
Number of columns, which is the number of items in the following lists. This will be the same for every row.

-> [Text]
List of column names. This will be the same for every row.

-> [Maybe Text]
List of column values, as returned by columnText.

-> IO ()

Statement management

withStatement

Arguments

:: Database
DB connection

-> Text
SQL statement

-> (Statement -> IO a)
User program

-> IO a

Bracket prepare and finalize. Useful for stepping multiple times through a Statement

prepare :: Database -> Text -> IO Statement

https://www.sqlite.org/c3ref/prepare.html

Unlike exec, prepare only executes the first statement, and ignores subsequent statements.

If the query string contains no SQL statements, this fails.

prepareUtf8 :: Database -> Utf8 -> IO Statement

https://www.sqlite.org/c3ref/prepare.html

It can help to avoid redundant Utf8 to Text conversion if you already have Utf8

If the query string contains no SQL statements, this fails.

step :: Statement -> IO StepResult

https://www.sqlite.org/c3ref/step.html

stepNoCB :: Statement -> IO StepResult

https://www.sqlite.org/c3ref/step.html

Faster step for statements that don't callback to Haskell functions (e.g. by using custom SQL functions).

reset :: Statement -> IO ()

https://www.sqlite.org/c3ref/reset.html

Note that in the C API, sqlite3_reset returns an error code if the most recent sqlite3_step indicated an error. We do not replicate that behavior here. reset never throws an exception.

finalize :: Statement -> IO ()

https://www.sqlite.org/c3ref/finalize.html

Like reset, finalize never throws an exception.

clearBindings :: Statement -> IO ()

https://www.sqlite.org/c3ref/clear_bindings.html

Set all parameters in the prepared statement to null.

Parameter and column information

bindParameterCount :: Statement -> IO ParamIndex

https://www.sqlite.org/c3ref/bind_parameter_count.html

This returns the index of the largest (rightmost) parameter. Note that this is not necessarily the number of parameters. If numbered parameters like ?5 are used, there may be gaps in the list.

See ParamIndex for more information.

bindParameterName :: Statement -> ParamIndex -> IO (Maybe Text)

https://www.sqlite.org/c3ref/bind_parameter_name.html

Return the N-th SQL parameter name.

Named parameters are returned as-is. E.g. ":v" is returned as Just ":v". Unnamed parameters, however, are converted to Nothing.

Note that the parameter index starts at 1, not 0.

columnCount :: Statement -> IO ColumnCount

https://www.sqlite.org/c3ref/column_count.html

columnName :: Statement -> ColumnIndex -> IO (Maybe Text)

https://www.sqlite.org/c3ref/column_name.html

Return the name of a result column. If the column index is out of range, return Nothing.

Binding values to a prepared statement

https://www.sqlite.org/c3ref/bind_blob.html

bindSQLData :: Statement -> ParamIndex -> SQLData -> IO ()

If the index is not between 1 and bindParameterCount inclusive, this fails with ErrorRange. Otherwise, it succeeds, even if the query skips this index by using numbered parameters.

Example:

> stmt <- prepare conn "SELECT ?1, ?3, ?5"
> bindSQLData stmt 1 (SQLInteger 1)
> bindSQLData stmt 2 (SQLInteger 2)
> bindSQLData stmt 6 (SQLInteger 6)
*** Exception: SQLite3 returned ErrorRange while attempting to perform bind int64.
> step stmt >> columns stmt
[SQLInteger 1,SQLNull,SQLNull]

bind :: Statement -> [SQLData] -> IO ()

Convenience function for binding values to all parameters. This will fail if the list has the wrong number of parameters.

bindNamed :: Statement -> [(Text, SQLData)] -> IO ()

Convenience function for binding named values to all parameters. This will fail if the list has the wrong number of parameters or if an unknown name is used.

Example:

stmt <- prepare conn "SELECT :foo + :bar"
bindNamed stmt [(":foo", SQLInteger 1), (":bar", SQLInteger 2)]

bindInt :: Statement -> ParamIndex -> Int -> IO ()

bindInt64 :: Statement -> ParamIndex -> Int64 -> IO ()

bindDouble :: Statement -> ParamIndex -> Double -> IO ()

bindText :: Statement -> ParamIndex -> Text -> IO ()

bindBlob :: Statement -> ParamIndex -> ByteString -> IO ()

bindZeroBlob :: Statement -> ParamIndex -> Int -> IO ()

bindNull :: Statement -> ParamIndex -> IO ()

Reading the result row

https://www.sqlite.org/c3ref/column_blob.html

Warning: column and columns will throw a DecodeError if any TEXT datum contains invalid UTF-8.

column :: Statement -> ColumnIndex -> IO SQLData

columns :: Statement -> IO [SQLData]

typedColumns :: Statement -> [Maybe ColumnType] -> IO [SQLData]

This avoids extra API calls using the list of column types. If passed types do not correspond to the actual types, the values will be converted according to the rules at https://www.sqlite.org/c3ref/column_blob.html. If the list contains more items that number of columns, the result is undefined.

columnType :: Statement -> ColumnIndex -> IO ColumnType

columnInt64 :: Statement -> ColumnIndex -> IO Int64

columnDouble :: Statement -> ColumnIndex -> IO Double

columnText :: Statement -> ColumnIndex -> IO Text

This will throw a DecodeError if the datum contains invalid UTF-8. If this behavior is undesirable, you can use columnText from Database.SQLite3.Direct, which does not perform conversion to Text.

columnBlob :: Statement -> ColumnIndex -> IO ByteString

Result statistics

lastInsertRowId :: Database -> IO Int64

https://www.sqlite.org/c3ref/last_insert_rowid.html

changes :: Database -> IO Int

https://www.sqlite.org/c3ref/changes.html

Return the number of rows that were changed, inserted, or deleted by the most recent INSERT, DELETE, or UPDATE statement.

Create custom SQL functions

createFunction

Arguments

:: Database

-> Text
Name of the function.

-> Maybe ArgCount
Number of arguments. Nothing means that the function accepts any number of arguments.

-> Bool
Is the function deterministic?

-> (FuncContext -> FuncArgs -> IO ())
Implementation of the function.

-> IO ()

https://sqlite.org/c3ref/create_function.html

Create a custom SQL function or redefine the behavior of an existing function. If the function is deterministic, i.e. if it always returns the same result given the same input, you can set the boolean flag to let sqlite perform additional optimizations.

createAggregate

Arguments

:: Database

-> Text
Name of the function.

-> Maybe ArgCount
Number of arguments.

-> a
Initial aggregate state.

-> (FuncContext -> FuncArgs -> a -> IO a)
Process one row and update the aggregate state.

-> (FuncContext -> a -> IO ())
Called after all rows have been processed. Can be used to construct the returned value from the aggregate state.

-> IO ()

Like createFunction except that it creates an aggregate function.

deleteFunction :: Database -> Text -> Maybe ArgCount -> IO ()

Delete an SQL function (scalar or aggregate).

Extract function arguments

funcArgCount :: FuncArgs -> ArgCount

funcArgType :: FuncArgs -> ArgIndex -> IO ColumnType

funcArgInt64 :: FuncArgs -> ArgIndex -> IO Int64

funcArgDouble :: FuncArgs -> ArgIndex -> IO Double

funcArgText :: FuncArgs -> ArgIndex -> IO Text

funcArgBlob :: FuncArgs -> ArgIndex -> IO ByteString

Set the result of a function

funcResultSQLData :: FuncContext -> SQLData -> IO ()

funcResultInt64 :: FuncContext -> Int64 -> IO ()

funcResultDouble :: FuncContext -> Double -> IO ()

funcResultText :: FuncContext -> Text -> IO ()

funcResultBlob :: FuncContext -> ByteString -> IO ()

funcResultZeroBlob :: FuncContext -> Int -> IO ()

funcResultNull :: FuncContext -> IO ()

getFuncContextDatabase :: FuncContext -> IO Database

https://www.sqlite.org/c3ref/context_db_handle.html

Create custom collations

createCollation

Arguments

:: Database

-> Text
Name of the collation.

-> (Text -> Text -> Ordering)
Comparison function.

-> IO ()

https://www.sqlite.org/c3ref/create_collation.html

deleteCollation :: Database -> Text -> IO ()

Delete a collation.

Interrupting a long-running query

interrupt :: Database -> IO ()

https://www.sqlite.org/c3ref/interrupt.html

Cause any pending operation on the Database handle to stop at its earliest opportunity. This simply sets a flag and returns immediately. It does not wait for the pending operation to finish.

You'll need to compile with -threaded for this to do any good. Without -threaded, FFI calls block the whole RTS, meaning interrupt would never run at the same time as step.

interruptibly :: Database -> IO a -> IO a

Make it possible to interrupt the given database operation with an asynchronous exception. This only works if the program is compiled with base >= 4.3 and -threaded.

It works by running the callback in a forked thread. If interrupted, it uses interrupt to try to stop the operation.

Incremental blob I/O

blobOpen

Arguments

:: Database

-> Text
The symbolic name of the database (e.g. "main").

-> Text
The table name.

-> Text
The column name.

-> Int64
The ROWID of the row.

-> Bool
Open the blob for read-write.

-> IO Blob

https://www.sqlite.org/c3ref/blob_open.html

Open a blob for incremental I/O.

blobClose :: Blob -> IO ()

https://www.sqlite.org/c3ref/blob_close.html

blobReopen

Arguments

:: Blob

-> Int64
The ROWID of the row.

-> IO ()

https://www.sqlite.org/c3ref/blob_reopen.html

blobBytes :: Blob -> IO Int

https://www.sqlite.org/c3ref/blob_bytes.html

blobRead

Arguments

:: Blob

-> Int
Number of bytes to read.

-> Int
Offset within the blob.

-> IO ByteString

https://www.sqlite.org/c3ref/blob_read.html

blobReadBuf :: Blob -> Ptr a -> Int -> Int -> IO ()

blobWrite

Arguments

:: Blob

-> ByteString

-> Int
Offset within the blob.

-> IO ()

https://www.sqlite.org/c3ref/blob_write.html

Online Backup API

https://www.sqlite.org/backup.html and https://www.sqlite.org/c3ref/backup_finish.html

backupInit

Arguments

:: Database
Destination database handle.

-> Text
Destination database name.

-> Database
Source database handle.

-> Text
Source database name.

-> IO Backup

backupFinish :: Backup -> IO ()

backupStep :: Backup -> Int -> IO BackupStepResult

backupRemaining :: Backup -> IO Int

https://www.sqlite.org/c3ref/backup_finish.html#sqlite3backupremaining

backupPagecount :: Backup -> IO Int

https://www.sqlite.org/c3ref/backup_finish.html#sqlite3backuppagecount

Types

data Database

data Statement

data SQLData

Constructors

SQLInteger !Int64

SQLFloat !Double

SQLText !Text

SQLBlob !ByteString

SQLNull

data SQLOpenFlag

These flags are used when using the open2 function. https://www.sqlite.org/c3ref/c_open_autoproxy.html

Constructors

SQLOpenReadOnly

SQLOpenReadWrite

SQLOpenCreate

SQLOpenDeleteOnClose

SQLOpenExclusive

SQLOpenAutoProxy

SQLOpenURI

SQLOpenMemory

SQLOpenMainDB

SQLOpenTempDB

SQLOpenTransientDB

SQLOpenMainJournal

SQLOpenTempJournal

SQLOpenSubJournal

SQLOpenMasterJournal

SQLOpenNoMutex

SQLOpenFullMutex

SQLOpenSharedCache

SQLOpenPrivateCache

SQLOpenWAL

SQLOpenNoFollow

SQLOpenExResCode

data SQLVFS

These VFS names are used when using the open2 function.

Constructors

SQLVFSDefault

SQLVFSUnix

SQLVFSUnixDotFile

SQLVFSUnixExcl

SQLVFSUnixNone

SQLVFSUnixNamedSem

SQLVFSCustom Text

data SQLError

Exception thrown when SQLite3 reports an error.

direct-sqlite may throw other types of exceptions if you misuse the API.

Constructors

SQLError

Fields

sqlError :: !Error

Error code returned by API call

sqlErrorDetails :: Text

Text describing the error

sqlErrorContext :: Text

Indicates what action produced this error, e.g. exec "SELECT * FROM foo"

data ColumnType

https://www.sqlite.org/c3ref/c_blob.html

Constructors

IntegerColumn

FloatColumn

TextColumn

BlobColumn

NullColumn

data FuncContext

The context in which a custom SQL function is executed.

data FuncArgs

The arguments of a custom SQL function.

data Blob

The type of blob handles used for incremental blob I/O

data Backup

A handle for an online backup process.

Results and errors

data StepResult

Constructors

Row

Done

data BackupStepResult

Constructors

BackupOK
There are still more pages to be copied.

BackupDone
All pages were successfully copied.

data Error

https://www.sqlite.org/c3ref/c_abort.html https://www.sqlite.org/c3ref/c_abort_rollback.html

Constructors

ErrorOK
Successful result

ErrorError
SQL error or missing database

ErrorInternal
Internal logic error in SQLite

ErrorPermission
Access permission denied

ErrorAbort
Callback routine requested an abort

ErrorBusy
The database file is locked

ErrorLocked
A table in the database is locked

ErrorNoMemory
A malloc() failed

ErrorReadOnly
Attempt to write a readonly database

ErrorInterrupt
Operation terminated by sqlite3_interrupt()

ErrorIO
Some kind of disk I/O error occurred

ErrorCorrupt
The database disk image is malformed

ErrorNotFound
Unknown opcode in sqlite3_file_control()

ErrorFull
Insertion failed because database is full

ErrorCan'tOpen
Unable to open the database file

ErrorProtocol
Database lock protocol error

ErrorEmpty
Database is empty

ErrorSchema
The database schema changed

ErrorTooBig
String or BLOB exceeds size limit

ErrorConstraint
Abort due to constraint violation

ErrorMismatch
Data type mismatch

ErrorMisuse
Library used incorrectly

ErrorNoLargeFileSupport
Uses OS features not supported on host

ErrorAuthorization
Authorization denied

ErrorFormat
Auxiliary database format error

ErrorRange
2nd parameter to sqlite3_bind out of range

ErrorNotADatabase
File opened that is not a database file

ErrorNotice
Notifications from sqlite3_log()

ErrorWarning
Warnings from sqlite3_log()

ErrorRow
sqlite3_step() has another row ready

ErrorDone
sqlite3_step() has finished executing

ErrorErrorMissingCollatingSquence

ErrorErrorRetry

ErrorErrorSnapshot

ErrorIORead

ErrorIOShortRead

ErrorIOWrite

ErrorIOFsync

ErrorIODirectoryFsync

ErrorIOTruncate

ErrorIOFstat

ErrorIOUnlock

ErrorIOReadLock

ErrorIOBlocked

ErrorIODelete

ErrorIONoMemory

ErrorIOAccess

ErrorIOCheckReservedLock

ErrorIOLock

ErrorIOClose

ErrorIODirectoryClose

ErrorIOShmOpen

ErrorIOShmSize

ErrorIOShmLock

ErrorIOShmMap

ErrorIOSeek

ErrorIODeleteNoEntity

ErrorIOMmap

ErrorIOGetTempPath

ErrorIOConvertedPath

ErrorIOVNode

ErrorIOAuth

ErrorIOBeginAtomic

ErrorIOCommitAtomic

ErrorIORollbackAtomic

ErrorIOData

ErrorIOCorruptFilesystem

ErrorLockedSharedCache

ErrorLockedVirtualTable

ErrorBusyRecovery

ErrorBusySnapshot

ErrorBusyTimeout

ErrorCan'tOpenNotTempDirectory

ErrorCan'tOpenIsDirectory

ErrorCan'tOpenFullPath

ErrorCan'tOpenConvertedPath

ErrorCan'tOpenDirtyWriteAheadLog

ErrorCan'tOpenSymlink

ErrorCorruptVirtualTable

ErrorCorruptSequence

ErrorCorruptIndex

ErrorReadOnlyRecovery

ErrorReadOnlyCan'tLock

ErrorReadOnlyRollback

ErrorReadOnlyDatabaseMoved

ErrorReadOnlyCan'tInit

ErrorReadOnlyDirectory

ErrorAbortRollback

ErrorConstraintCheck

ErrorConstraintCommitHook

ErrorConstraintForeignKey

ErrorConstraintFunction

ErrorConstraintNotNull

ErrorConstraintPrimaryKey

ErrorConstraintTrigger

ErrorConstraintUnique

ErrorConstraintVirtualTable

ErrorConstraintRowId

ErrorConstraintPinned

ErrorConstraintDataType

ErrorNoticeRecoverWriteAheadLog

ErrorNoticeRecoverRollback

ErrorWarningAutoIndex

ErrorAuthUser

ErrorOkLoadPermanently

Special integers

newtype ParamIndex

Index of a parameter in a parameterized query. Parameter indices start from 1.

When a query is prepared, SQLite allocates an array indexed from 1 to the highest parameter index. For example:

>Right stmt <- prepare conn "SELECT ?1, ?5, ?3, ?"
>bindParameterCount stmt
ParamIndex 6

This will allocate an array indexed from 1 to 6 (? takes the highest preceding index plus one). The array is initialized with null values. When you bind a parameter with bindSQLData, it assigns a new value to one of these indices.

See https://www.sqlite.org/lang_expr.html#varparam for the syntax of parameter placeholders, and how parameter indices are assigned.

Constructors

ParamIndex Int

newtype ColumnIndex

Index of a column in a result set. Column indices start from 0.

Constructors

ColumnIndex Int

type ColumnCount = ColumnIndex

Number of columns in a result set.

newtype ArgCount

Number of arguments of a user defined SQL function.

Constructors

ArgCount Int

type ArgIndex = ArgCount

Index of an argument to a custom function. Indices start from 0.