Database.SQLite3.Direct

Description

This API is a slightly lower-level version of Database.SQLite3. Namely:

It returns errors instead of throwing them.

It only uses cheap conversions. None of these bindings convert from String or Text.

Connection management

open :: Utf8 -> IO (Either (Error, Utf8) Database)

http://www.sqlite.org/c3ref/open.html

close :: Database -> IO (Either Error ())

http://www.sqlite.org/c3ref/close.html

errmsg :: Database -> IO Utf8

http://www.sqlite.org/c3ref/errcode.html

setTrace :: Database -> Maybe (Utf8 -> IO ()) -> IO ()

http://www.sqlite.org/c3ref/profile.html

Enable/disable tracing of SQL execution. Tracing can be disabled by setting Nothing as the logger callback.

Warning: If the logger callback throws an exception, your whole program will crash. Enable only for debugging!

getAutoCommit :: Database -> IO Bool

http://www.sqlite.org/c3ref/get_autocommit.html

Return True if the connection is in autocommit mode, or False if a transaction started with BEGIN is still active.

Be warned that some errors roll back the transaction automatically, and that ROLLBACK will throw an error if no transaction is active. Use getAutoCommit to avoid such an error:

  autocommit <- getAutoCommit conn
  when (not autocommit) $
      exec conn "ROLLBACK"

Simple query execution

http://sqlite.org/c3ref/exec.html

exec :: Database -> Utf8 -> IO (Either (Error, Utf8) ())

execWithCallback :: Database -> Utf8 -> ExecCallback -> IO (Either (Error, Utf8) ())

Like exec, but invoke the callback for each result row.

If the callback throws an exception, it will be rethrown by execWithCallback.

type ExecCallback

Arguments

= ColumnCount
Number of columns, which is the number of items in the following lists. This will be the same for every row.

-> [Utf8]
List of column names. This will be the same for every row.

-> [Maybe Utf8]
List of column values, as returned by columnText.

-> IO ()

Statement management

prepare :: Database -> Utf8 -> IO (Either Error (Maybe Statement))

http://www.sqlite.org/c3ref/prepare.html

If the query contains no SQL statements, this returns Right Nothing.

getStatementDatabase :: Statement -> IO Database

http://www.sqlite.org/c3ref/db_handle.html

step :: Statement -> IO (Either Error StepResult)

http://www.sqlite.org/c3ref/step.html

reset :: Statement -> IO (Either Error ())

http://www.sqlite.org/c3ref/reset.html

Warning:

If the most recent step call failed, this will return the corresponding error.

This does not reset the bindings on a prepared statement. Use clearBindings to do that.

finalize :: Statement -> IO (Either Error ())

http://www.sqlite.org/c3ref/finalize.html

Warning: If the most recent step call failed, this will return the corresponding error.

clearBindings :: Statement -> IO ()

http://www.sqlite.org/c3ref/clear_bindings.html

Set all parameters in the prepared statement to null.

statementSql :: Statement -> IO (Maybe Utf8)

http://www.sqlite.org/c3ref/sql.html

Return a copy of the original SQL text used to compile the statement.

Parameter and column information

bindParameterCount :: Statement -> IO ParamIndex

http://www.sqlite.org/c3ref/bind_parameter_count.html

This returns the index of the largest (rightmost) parameter. Note that this is not necessarily the number of parameters. If numbered parameters like ?5 are used, there may be gaps in the list.

See ParamIndex for more information.

bindParameterName :: Statement -> ParamIndex -> IO (Maybe Utf8)

http://www.sqlite.org/c3ref/bind_parameter_name.html

columnCount :: Statement -> IO ColumnCount

http://www.sqlite.org/c3ref/column_count.html

columnName :: Statement -> ColumnIndex -> IO (Maybe Utf8)

http://www.sqlite.org/c3ref/column_name.html

Binding values to a prepared statement

http://www.sqlite.org/c3ref/bind_blob.html

bindInt64 :: Statement -> ParamIndex -> Int64 -> IO (Either Error ())

bindDouble :: Statement -> ParamIndex -> Double -> IO (Either Error ())

bindText :: Statement -> ParamIndex -> Utf8 -> IO (Either Error ())

bindBlob :: Statement -> ParamIndex -> ByteString -> IO (Either Error ())

bindNull :: Statement -> ParamIndex -> IO (Either Error ())

Reading the result row

http://www.sqlite.org/c3ref/column_blob.html

columnType :: Statement -> ColumnIndex -> IO ColumnType

columnInt64 :: Statement -> ColumnIndex -> IO Int64

columnDouble :: Statement -> ColumnIndex -> IO Double

columnText :: Statement -> ColumnIndex -> IO Utf8

columnBlob :: Statement -> ColumnIndex -> IO ByteString

Result statistics

lastInsertRowId :: Database -> IO Int64

http://www.sqlite.org/c3ref/last_insert_rowid.html

changes :: Database -> IO Int

http://www.sqlite.org/c3ref/changes.html

Return the number of rows that were changed, inserted, or deleted by the most recent INSERT, DELETE, or UPDATE statement.

totalChanges :: Database -> IO Int

http://www.sqlite.org/c3ref/total_changes.html

Return the total number of row changes caused by INSERT, DELETE, or UPDATE statements since the Database was opened.

Interrupting a long-running query

interrupt :: Database -> IO ()

http://www.sqlite.org/c3ref/interrupt.html

Cause any pending operation on the Database handle to stop at its earliest opportunity. This simply sets a flag and returns immediately. It does not wait for the pending operation to finish.

You'll need to compile with -threaded for this to do any good. Without -threaded, FFI calls block the whole RTS, meaning interrupt would never run at the same time as step.

Types

newtype Database

Constructors

Database (Ptr CDatabase)

newtype Statement

Constructors

Statement (Ptr CStatement)

data ColumnType

Constructors

IntegerColumn

FloatColumn

TextColumn

BlobColumn

NullColumn

Results and errors

data StepResult

Constructors

Row

Done

data Error

Constructors

ErrorOK
Successful result

ErrorError
SQL error or missing database

ErrorInternal
Internal logic error in SQLite

ErrorPermission
Access permission denied

ErrorAbort
Callback routine requested an abort

ErrorBusy
The database file is locked

ErrorLocked
A table in the database is locked

ErrorNoMemory
A malloc() failed

ErrorReadOnly
Attempt to write a readonly database

ErrorInterrupt
Operation terminated by sqlite3_interrupt()

ErrorIO
Some kind of disk I/O error occurred

ErrorCorrupt
The database disk image is malformed

ErrorNotFound
Unknown opcode in sqlite3_file_control()

ErrorFull
Insertion failed because database is full

ErrorCan'tOpen
Unable to open the database file

ErrorProtocol
Database lock protocol error

ErrorEmpty
Database is empty

ErrorSchema
The database schema changed

ErrorTooBig
String or BLOB exceeds size limit

ErrorConstraint
Abort due to constraint violation

ErrorMismatch
Data type mismatch

ErrorMisuse
Library used incorrectly

ErrorNoLargeFileSupport
Uses OS features not supported on host

ErrorAuthorization
Authorization denied

ErrorFormat
Auxiliary database format error

ErrorRange
2nd parameter to sqlite3_bind out of range

ErrorNotADatabase
File opened that is not a database file

ErrorRow
sqlite3_step() has another row ready

ErrorDone
sqlite3_step() has finished executing

Special types

newtype Utf8

A ByteString containing UTF8-encoded text with no NUL characters.

Constructors

Utf8 ByteString

newtype ParamIndex

Index of a parameter in a parameterized query. Parameter indices start from 1.

When a query is prepared, SQLite allocates an array indexed from 1 to the highest parameter index. For example:

>Right stmt <- prepare conn "SELECT ?1, ?5, ?3, ?"
>bindParameterCount stmt
ParamIndex 6

This will allocate an array indexed from 1 to 6 (? takes the highest preceding index plus one). The array is initialized with null values. When you bind a parameter with bindSQLData, it assigns a new value to one of these indices.

See http://www.sqlite.org/lang_expr.html#varparam for the syntax of parameter placeholders, and how parameter indices are assigned.

Constructors

ParamIndex Int

newtype ColumnIndex

Index of a column in a result set. Column indices start from 0.

Constructors

ColumnIndex Int

type ColumnCount = ColumnIndex

Number of columns in a result set.